requirements.txt
easydict==1.9
matplotlib==3.0.3
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn==0.0
mxnet-cu90==1.4.0
